+++ lib/main.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:hoowlib/providers/appsettings_provider.dart';
+import 'package:hoowlib/pages/dev_page.dart';
 import 'models/book.dart';
 import 'pages/library_page.dart';
 import 'pages/calendar_page.dart';
@@ -6,6 +8,8 @@ import 'pages/setting_page.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'l10n/app_localizations.dart';
 import 'db/database_helper.dart';
+import 'package:provider/provider.dart';
+import 'providers/books_provider.dart';
 import 'dart:ui' as ui;
 
 void main() {
@@ -42,7 +46,15 @@ class _MainAppState extends State<MainApp> {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return MultiProvider(providers: [
+      ChangeNotifierProvider(
+        create: (context) => AppsettingsProvider()
+        ),
+      ChangeNotifierProvider(
+        create: (context) => BooksProvider()
+        ),
+    ],
+    child: MaterialApp(
       debugShowCheckedModeBanner: false,
       locale: _locale,
       localizationsDelegates: const [
@@ -56,7 +68,7 @@ class _MainAppState extends State<MainApp> {
         Locale('de', ''),
       ],
       home: MainNavigation(onLocaleChanged: _changeLocale),
-    );
+    ),);
   }
 }
 
@@ -71,6 +83,7 @@ class MainNavigation extends StatefulWidget {
 class _MainNavigationState extends State<MainNavigation> {
   int _selectedIndex = 0;
   List<Book> _books = [];
+  bool _devMode = false;
 
   final GlobalKey<LibraryPageState> _libraryPageKey = GlobalKey<LibraryPageState>();
 
@@ -78,6 +91,12 @@ class _MainNavigationState extends State<MainNavigation> {
   void initState() {
     super.initState();
     _loadBooks();
+  }
+
+  void _changeDevMode(bool newDevMode) {
+    setState(() {
+      _devMode = newDevMode;
+    });
   }
 
   Future<void> _loadBooks() async {
@@ -90,7 +109,8 @@ class _MainNavigationState extends State<MainNavigation> {
   List<Widget> get _pages => [
         LibraryPage(key: _libraryPageKey, books: _books),
         const CalendarPage(),
-        SettingPage(onLocaleChanged: widget.onLocaleChanged),
+        SettingPage(onLocaleChanged: widget.onLocaleChanged, onDevModeChanged: _changeDevMode, devModeParameter: _devMode,),
+        DevPage(),
       ];
 
   void _onItemTapped(int index) {
@@ -101,16 +121,9 @@ class _MainNavigationState extends State<MainNavigation> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: _pages[_selectedIndex],
-      bottomNavigationBar: Container(
-        constraints: BoxConstraints(minHeight: 70, maxHeight: 90),
-        child: BottomNavigationBar(
-          type: BottomNavigationBarType.fixed,
-          selectedFontSize: 12,
-          unselectedFontSize: 12,
-          items: <BottomNavigationBarItem>[
-            BottomNavigationBarItem(
+
+    List<BottomNavigationBarItem> bottom = [
+      BottomNavigationBarItem(
               icon: Icon(Icons.home, size: 24),
               label: AppLocalizations.of(context)!.myLibrary,
             ),
@@ -122,7 +135,24 @@ class _MainNavigationState extends State<MainNavigation> {
               icon: Icon(Icons.settings, size: 24),
               label: AppLocalizations.of(context)!.settings,
             ),
-          ],
+    ];
+
+    if(Provider.of<AppsettingsProvider>(context).devMode){
+      bottom.add(BottomNavigationBarItem(
+              icon: Icon(Icons.developer_mode, size: 24,),
+              label: "dev",
+            ));
+    }
+
+    return Scaffold(
+      body: _pages[_selectedIndex],
+      bottomNavigationBar: Container(
+        constraints: BoxConstraints(minHeight: 70, maxHeight: 90),
+        child: BottomNavigationBar(
+          type: BottomNavigationBarType.fixed,
+          selectedFontSize: 12,
+          unselectedFontSize: 12,
+          items: bottom,
           currentIndex: _selectedIndex,
           onTap: _onItemTapped,
         ),