+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.8.1] - open
+## [0.9.0] - open
+- add developer mode setting
+- add developer page
+- update library page with provider
+- update dev page
+
+## [0.8.1] - 2025-11-20
 - library sorting to case-insensitive
 - update appbar ui on calendar and settings
 - update ui detail book sheet